paper_reproducibility/requirements.txt
absl-py==2.3.1
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.4
colorlog==4.7.2
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
grpcio==1.73.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Markdown==3.8.2
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
narwhals==1.47.0
networkx==3.5
numpy==1.26.4
packaging==25.0
pandas==2.3.1
plotly==6.2.0
protobuf==6.31.1
psutil==7.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
ray==2.6.3
recbole==1.2.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
texttable==1.7.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
torch==2.5.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
